refactor: moved test files and ChangeLogConfig
